Move shared clients to `pkg/kubernetes`
